move search results header to dedicated file. Fix sorting bug
